pallet_bridge_grandpa::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

BestFinalized

Hash of the best finalized header.

CurrentAuthoritySet

The current GRANDPA Authority set.

FreeHeadersRemaining

Number of free header submissions that we may yet accept in the current block.

ImportedHashes

A ring buffer of imported hashes. Ordered by the insertion time.

ImportedHashesPointer

Current ring buffer position.

ImportedHeaders

Relevant fields of imported headers.

InitialHash

Hash of the header used to bootstrap the pallet.

PalletOperatingMode

The current operating mode of the pallet.

PalletOwner

Optional pallet owner.