Trait pallet_bridge_grandpa::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type BridgedChain: Chain;
    type MaxRequests: Get<u32>;
    type HeadersToKeep: Get<u32>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type BridgedChain: Chain

The chain we are bridging to here.

type MaxRequests: Get<u32>

The upper bound on the number of requests allowed by the pallet.

A request refers to an action which writes a header to storage.

Once this bound is reached the pallet will not allow any dispatchables to be called until the request count has decreased.

type HeadersToKeep: Get<u32>

Maximal number of finalized headers to keep in the storage.

The setting is there to prevent growing the on-chain state indefinitely. Note the setting does not relate to block numbers - we will simply keep as much items in the storage, so it doesn’t guarantee any fixed timeframe for finality headers.

type WeightInfo: WeightInfo

Weights gathered through benchmarking.