Type Alias pallet_bridge_grandpa::GrandpaChainHeaders

pub type GrandpaChainHeaders<T, I> = Pallet<T, I>;

Bridge GRANDPA pallet as header chain.

Aliased Type

struct GrandpaChainHeaders<T, I>(/* private fields */);

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn submit_finality_proof( origin: OriginFor<T>, finality_target: Box<BridgedHeader<T, I>>, justification: GrandpaJustification<BridgedHeader<T, I>> ) -> DispatchResultWithPostInfo

Verify a target header is finalized according to the given finality proof.

It will use the underlying storage pallet to fetch information about the current authorities and best finalized header in order to verify that the header is finalized.

If successful in verification, it will write the target header to the underlying storage pallet.

The call fails if:

the pallet is halted;

the pallet knows better header than the finality_target;

verification is not optimized or invalid;

header contains forced authorities set change or change with non-zero delay.

pub fn initialize( origin: OriginFor<T>, init_data: InitializationData<BridgedHeader<T, I>> ) -> DispatchResultWithPostInfo

Bootstrap the bridge pallet with an initial header and authority set from which to sync.

The initial configuration provided does not need to be the genesis header of the bridged chain, it can be any arbitrary header. You can also provide the next scheduled set change if it is already know.

This function is only allowed to be called from a trusted origin and writes to storage with practically no checks in terms of the validity of the data. It is important that you ensure that valid data is being passed in.

pub fn set_owner( origin: OriginFor<T>, new_owner: Option<T::AccountId> ) -> DispatchResult

Change PalletOwner.

May only be called either by root, or by PalletOwner.

pub fn set_operating_mode( origin: OriginFor<T>, operating_mode: BasicOperatingMode ) -> DispatchResult

Halt or resume all pallet operations.

May only be called either by root, or by PalletOwner.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn free_mandatory_headers_remaining() -> u32

An auto-generated getter for FreeMandatoryHeadersRemaining.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn best_finalized() -> Option<BridgedBlockId<T, I>>

An auto-generated getter for BestFinalized.

impl<T: Config<I>, I: 'static> Pallet<T, I>where <T as Config>::RuntimeEvent: TryInto<Event<T, I>>,

pub fn synced_headers_grandpa_info( ) -> Vec<StoredHeaderGrandpaInfo<BridgedHeader<T, I>>>

Get the GRANDPA justifications accepted in the current block.