Crate pallet_bridge_grandpa

Substrate GRANDPA Pallet

This pallet is an on-chain GRANDPA light client for Substrate based chains.

This pallet achieves this by trustlessly verifying GRANDPA finality proofs on-chain. Once verified, finalized headers are stored in the pallet, thereby creating a sparse header chain. This sparse header chain can be used as a source of truth for other higher-level applications.

The pallet is responsible for tracking GRANDPA validator set hand-offs. We only import headers with justifications signed by the current validator set we know of. The header is inspected for a ScheduledChanges digest item, which is then used to update to next validator set.

Since this pallet only tracks finalized headers it does not deal with forks. Forks can only occur if the GRANDPA validator set on the bridged chain is either colluding or there is a severe bug causing resulting in an equivocation. Such events are outside of the scope of this pallet. Shall the fork occur on the bridged chain governance intervention will be required to re-initialize the bridge and track the right fork.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Pallet containing weights for this pallet. Autogenerated weights for pallet_bridge_grandpa

Type Definitions

BridgedBlockHash

Block hash of the bridged chain.

BridgedBlockHasher

Hasher of the bridged chain.

BridgedBlockNumber

Block number of the bridged chain.

BridgedHeader

Header of the bridged chain.