Trait pallet_bridge_grandpa::CallSubType

pub trait CallSubType<T: Config<I, RuntimeCall = Self>, I: 'static>: IsSubType<CallableCallFor<Pallet<T, I>, T>> {
    // Provided methods
    fn submit_finality_proof_info(
        &self
    ) -> Option<SubmitFinalityProofInfo<BridgedBlockNumber<T, I>>> { ... }
    fn check_obsolete_submit_finality_proof(&self) -> TransactionValidity
       where Self: Sized { ... }
}

Trait representing a call that is a sub type of this pallet’s call.

Provided Methods

fn submit_finality_proof_info( &self ) -> Option<SubmitFinalityProofInfo<BridgedBlockNumber<T, I>>>

Extract finality proof info from a runtime call.

fn check_obsolete_submit_finality_proof(&self) -> TransactionValiditywhere Self: Sized,

Validate Grandpa headers in order to avoid “mining” transactions that provide outdated bridged chain headers. Without this validation, even honest relayers may lose their funds if there are multiple relays running and submitting the same information.